Creates a children API module and moves one of the functions to it
